feat: WIP attachment component for sm
