+++ frontend/src/casedata/components/errand/tabs/attachments/casedata-attachments.component.tsx
@@ -331,7 +331,6 @@ export const CasedataAttachments: React.FC = () => {
         newFiles={newFiles}
         attachmentTypeExists={attachmentTypeExists}
         errand={errand}
-        municipalityId={municipalityId}
         saveErrand={saveErrand}
         setErrand={setErrand}
         closeHandler={closeHandler}
